% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_url.R
\name{apply_additional_endpoints}
\alias{apply_additional_endpoints}
\title{Append an additional endpoint to a given URL}
\usage{
apply_additional_endpoints(url, additional_endpoint)
}
\arguments{
\item{url}{A character string representing the base URL to which the additional
endpoint will be appended.}

\item{additional_endpoint}{A character string representing the additional
endpoint to be appended to the base URL.}
}
\value{
A character string representing the new URL with the additional
endpoint appended.
}
\description{
This function appends an additional endpoint to a given URL while ensuring
that there are no trailing slashes in the base URL.
}
