% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{construct_basic_response_msg}
\alias{construct_basic_response_msg}
\title{Construct a basic HTTP response message}
\usage{
construct_basic_response_msg(func_name, response)
}
\arguments{
\item{func_name}{A character string representing the name of the
function that made the request.}

\item{response}{An object of class \code{response} from the \code{httr} package.}
}
\value{
A character string representing the HTTP response message.
}
\description{
This function constructs a basic HTTP response message by combining
the name of the function that made the request and the HTTP status
code returned by the server.
}
\examples{
\dontrun{
construct_basic_response_msg("get_data", response)
}

}
