% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{extend_msg_error}
\alias{extend_msg_error}
\title{Extend an error message with additional details}
\usage{
extend_msg_error(init_msg, response, url)
}
\arguments{
\item{init_msg}{A character string representing the initial error message.}

\item{response}{An object of class \code{response} from the \code{httr} package.}

\item{url}{A character string representing the URL that was requested.}
}
\value{
A character string representing the extended error message.
}
\description{
This function takes an initial error message, an HTTP response object, and a URL, and extends the error message with additional details about the error.
}
\examples{
\dontrun{
extend_msg_error(
  construct_basic_response_msg("func", response),
  response, "https://example.com/data"
)
}

}
