% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data}
\alias{get_data}
\title{Get data from a REST API with pagination}
\usage{
get_data(
  url,
  token,
  pagination_endpoint = database_info$endpoints$pagination_info,
  count_endpoint = database_info$endpoints$row_count
)
}
\arguments{
\item{url}{URL of the REST API}

\item{token}{The authentication token to use when accessing the endpoint}

\item{pagination_endpoint}{Endpoint for pagination}

\item{count_endpoint}{Endpoint for counting total records}
}
\value{
A tibble containing the data retrieved from the API
}
\description{
This function retrieves data from a REST API that uses pagination.
}
\examples{
\dontrun{
get_data("https://example.com/api", "1234")
}

}
\seealso{
\code{\link{get_pagination_info}}, \code{\link{get_total_records}}
}
