% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_field_info}
\alias{get_field_info}
\title{Retrieve the fields of a given table from a fields endpoint}
\usage{
get_field_info(url, token, fields_endpoint = "fields/")
}
\arguments{
\item{url}{The base URL of the endpoint, such as
"http://127.0.0.1:8000/api/v1/genes/".}

\item{token}{Your API authentication token, if required.}

\item{fields_endpoint}{The endpoint for retrieving the fields,
by default set to "fields/".}
}
\value{
A list containing the readable fields, writable fields, and
automatically generated fields.
}
\description{
This function retrieves the fields of a given table from a fields
endpoint and returns them as a list containing readable fields,
writable fields, and automatically generated fields.
}
\examples{
\dontrun{
# Get the fields of a given table from a fields endpoint
url <- "https://example.com/api/genes/"
token <- "my_token"
table_fields <- get_field_info(url, token)
}

}
