% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_paginated_data}
\alias{get_paginated_data}
\title{Retrieve Paginated Data From a REST API}
\usage{
get_paginated_data(init_url, token)
}
\arguments{
\item{init_url}{The initial URL from which to start retrieving data.}

\item{token}{The authorization token for the API.}
}
\value{
A tibble containing all data retrieved from the API.
}
\description{
This function retrieves all data from a paginated REST API. It will continue
to request data until all pages have been retrieved.
}
\examples{
\dontrun{
  get_paginated_data("https://api.example.com/data",
                         "your_token")
}

}
