% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_pagination_info}
\alias{get_pagination_info}
\title{Get pagination information from a REST_FRAMEWORK endpoint}
\usage{
get_pagination_info(
  url,
  token,
  pagination_endpoint = database_info$endpoints$pagination_info
)
}
\arguments{
\item{url}{A character string specifying the view against which to send
a request. This should be constructed using database_info -- see the
example.}

\item{token}{The authentication token to use when accessing the endpoint}

\item{pagination_endpoint}{By default, this is set to
database_info$endpoints$pagination_info and will be appended appropriately
to the url. For example, if you are interested in retrieving the
pagination settings for the following table:
http://127.0.0.1:8000/api/v1/genes/, then you would pass this url, with
or without the trailing /, and the pagination_endpoint, if it is not
provided in the database_info. The URL which will be queried will be
something along the lines of:
http://127.0.0.1:8000/api/v1/genes/pagination_info/. Of course, this
endpoint must exist for this to work.}
}
\value{
A list with the default page size and page size limit
}
\description{
This function retrieves the pagination information for a given
REST_FRAMEWORK endpoint. It returns a list with the default page
size and page size limit.
}
\examples{
\dontrun{
token <- get_user_auth_token(
  paste0(
    database_info$base_url,
    database_info$endpoints$auth_token
  ),
  "username",
  "password"
)

url <- paste0(database_info$base_url, database_info$endpoints$chrmap)
# Retrieve the pagination information
pagination_info <- get_pagination_info(url, token)

print(pagination_info)
}

}
\seealso{
\code{\link{get_user_auth_token}},
\code{\link{database_info}}
}
