% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_record_count}
\alias{get_record_count}
\title{Retrieve the total number of records from a count endpoint}
\usage{
get_record_count(
  url,
  token,
  count_endpoint = database_info$endpoints$row_count
)
}
\arguments{
\item{url}{A character string specifying the view against which to send
a request. This should be constructed using database_info -- see the
example.}

\item{token}{The authentication token to use when accessing the endpoint}

\item{count_endpoint}{By default, this is set to
database_info$endpoints$row_count and will be appended appropriately
to the url. For example, if you are interested in retrieving the
row count for the following table:
http://127.0.0.1:8000/api/v1/genes/, then you would pass this url, with
or without the trailing /, and the count_endpoint, if it is not
provided in the database_info. The URL which will be queried will be
something along the lines of:
http://127.0.0.1:8000/api/v1/genes/count/. Of course, this
endpoint must exist for this to work.}
}
\value{
The total number of records as a numeric value.
}
\description{
This function retrieves the total number of records from a count
endpoint and returns it as a numeric value.
}
\examples{
\dontrun{
# Get the total number of records from a count endpoint
count_url <- "https://example.com/api/count"
token <- "my_token"
total_records <- get_record_count(count_url, token)
}

}
\seealso{
\code{\link{database_info}},
\code{\link{get_user_auth_token}}
}
