% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_table_by_page}
\alias{get_table_by_page}
\title{Create a function to return a function to get paginated results}
\usage{
get_table_by_page(url, token, page_size)
}
\arguments{
\item{url}{The URL of the endpoint to retrieve data from}

\item{token}{The authentication token to use for the GET request}

\item{page_size}{The number of records to retrieve per page}
}
\value{
A function that can be called with a start index to retrieve
a page of results
}
\description{
This function returns a function that can be called to make a GET request to
retrieve data from a paginated API endpoint.
}
\examples{
\dontrun{
get_by_page <- get_table_by_page(
  "http://example.com/api/endpoint",
  "my_auth_token", 10
)

get_by_page(1)
}

}
