% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_total_records}
\alias{get_total_records}
\title{Retrieve the total number of records from a count endpoint}
\usage{
get_total_records(count_url, token)
}
\arguments{
\item{count_url}{The URL of the count endpoint.}

\item{token}{The authentication token to use when accessing the endpoint}
}
\value{
The total number of records as a numeric value.
}
\description{
This function retrieves the total number of records from a count
endpoint and returns it as a numeric value.
}
\examples{
\dontrun{
# Get the total number of records from a count endpoint
count_url <- "https://example.com/api/count"
token <- "my_token"
total_records <- get_total_records(count_url, token)
}
}
\seealso{
\code{\link{get_pagination_info}}, \code{\link{get_user_auth_token}}
}
