% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{post_json}
\alias{post_json}
\title{a helper for send() function for json data}
\usage{
post_json(df, url, token)
}
\arguments{
\item{df}{a dataframe with the column \code{id}}

\item{url}{the url to the table endpoint -- the \code{id} is added in
this function}

\item{token}{authorization token}
}
\description{
a helper for send() function for json data
}
