% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{post_postgres_csv}
\alias{post_postgres_csv}
\title{a helper for send() function for csv data}
\usage{
post_postgres_csv(
  df,
  url,
  token,
  csv_postgres_endpoint = "upload-csv-postgres/"
)
}
\arguments{
\item{df}{a dataframe with the column \code{id}}

\item{url}{the url to the table endpoint -- the \code{id} is added in
this function}

\item{token}{authorization token}

\item{csv_postgres_endpoint}{an additional string to append to the
url which will reach an endpoint for a given view that is configured
for bulk uploads of csv files, eg 'upload-csv/'. NOTE this also sets
the upload data type to csv currently}
}
\description{
a helper for send() function for csv data
}
