% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve from a REST API with pagination}
\usage{
retrieve(url, token, ...)
}
\arguments{
\item{url}{A character string specifying the view against which to send
a request. This should be constructed using database_info -- see the
example.}

\item{token}{The authentication token to use when accessing the endpoint}

\item{...}{
  Arguments passed on to \code{\link[=apply_filters_to_url]{apply_filters_to_url}}
  \describe{
    \item{\code{filter_list}}{A named list of filters to apply to the URL.}
  }}
}
\value{
A tibble containing the data retrieved from the API
}
\description{
This function retrieves data from a REST API that uses pagination.
}
\examples{
\dontrun{
retrieve("https://example.com/api", "1234")
}

}
\seealso{
\code{\link{get_pagination_info}}, \code{\link{get_record_count}}
}
