% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send}
\alias{send}
\title{Send a dataframe to a server endpoint}
\usage{
send(df, url, token)
}
\arguments{
\item{df}{A dataframe to send to the server endpoint.}

\item{url}{A character string specifying the view against which to send
a request. This should be constructed using database_info -- see the
example.}

\item{token}{The authentication token to use when accessing the endpoint}
}
\value{
Nothing is returned explicitly; the function logs the HTTP
response status and message instead.
}
\description{
This function sends a dataframe to a server endpoint using the POST method
and the JSON format.
}
\examples{
\dontrun{
labretriever::send(df, "http://example.com/api/myendpoint", "my_token")
}
}
