% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_info}
\alias{database_info}
\title{Database Information for the Yeast Compendium API}
\format{
A list containing the following elements:
\describe{
\item{base_url}{The base URL for the API (a character string).}
\item{endpoints}{
A list of API endpoints (character strings) for various resources:
\describe{
\item{auth_token}{Authentication token endpoint.}
\item{chrmap}{Chromosome map endpoint.}
\item{genes}{Genes endpoint.}
\item{genes_with_effects}{Genes with effects endpoint.}
\item{promoterregions}{Promoter regions endpoint.}
\item{promoterregions_targets}{Promoter regions targets endpoint.}
\item{harbisonchip}{Harbison ChIP endpoint.}
\item{harbisonchip_with_annote}{Harbison ChIP with
annotations endpoint.}
\item{kemmerentfko}{Kemmeren TF knockout endpoint.}
\item{mcisaaczev}{McIsaac ZEV endpoint.}
\item{background}{Background endpoint.}
\item{cctf}{CC TF endpoint.}
\item{cctf_tf_list}{a dataframe with columns tf_id, tf_locus_tag,
tf_gene which describes the TFs which have been interrogated by
calling cards}
\item{ccexperiment}{CC experiment endpoint.}
\item{hops}{HOPS endpoint.}
\item{hopsreplicatesig}{HOPS replicate significance endpoint.}
\item{hopsreplicatesig_with_annote}{HOPS replicate significance
with annotations endpoint.}
\item{qcmetrics}{QC metrics endpoint.}
\item{qcmanualreview}{QC manual review endpoint.}
\item{qcreview}{This is the endpoint for the summarized QC view with which most users will want to interactive}
\item{qcr1tor2}{QC R1 to R2 endpoint.}
\item{qcr2tor1}{QC R2 to R1 endpoint.}
\item{qctftotransposon}{QC TF to transposon endpoint.}
\item{expression}{Expression endpoint.}
\item{pagination_info}{Pagination information endpoint.}
\item{row_count}{Row count endpoint.}
}
}
}
}
\usage{
database_info
}
\description{
A list containing the base URL and various API endpoints to interact
with the Yeast Compendium API.
}
\examples{
# Accessing the base_url from the database_info object
database_info$base_url
# Accessing the genes endpoint from the database_info object
database_info$endpoints$genes

}
\keyword{datasets}
