% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve from a REST API with pagination}
\usage{
retrieve(url, token, retrieve_all = TRUE, ...)
}
\arguments{
\item{url}{The base URL of the endpoint, such as
"http://127.0.0.1:8000/api/v1/genes/".}

\item{token}{Your API authentication token, if required.}

\item{retrieve_all}{Boolean. Default TRUE. Set to FALSE to return
only a single page, if the data is paginated}

\item{...}{
  Arguments passed on to \code{\link[=apply_filters_to_url]{apply_filters_to_url}}
  \describe{
    \item{\code{filter_list}}{A named list of filters to apply to the URL.}
  }}
}
\value{
A tibble containing the data retrieved from the API
}
\description{
This function retrieves data from a REST API that uses pagination.
}
\examples{
\dontrun{
retrieve("https://example.com/api", "1234")
}

}
\seealso{
\code{\link{get_pagination_info}}, \code{\link{get_record_count}}
}
