% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send}
\alias{send}
\title{Send a dataframe to a server endpoint}
\usage{
send(df, url, token, strict = TRUE, update = FALSE)
}
\arguments{
\item{df}{A dataframe to send to the server endpoint.}

\item{url}{A character string specifying the view against which to send
a request. This should be constructed using database_info -- see the
example.}

\item{token}{The authentication token to use when accessing the endpoint}

\item{strict}{Ensure that writable columns in the database table are
present in input df. the Defaults to TRUE.}

\item{update}{Set to TRUE if you are updating a table. This will turn off
the strict check -- only the 'id' column is required.
Additional columns will be those updated, according to the 'id'. This
will be used to append to the url}
}
\value{
Nothing is returned explicitly; the function logs the HTTP
response status and message instead.
}
\description{
This function sends a dataframe to a server endpoint using the POST method
and the JSON format.
}
\examples{
\dontrun{
labretriever::send(df, "http://example.com/api/myendpoint", "my_token")
}
}
